+++ pipe_operator/python_flow/__init__.py
@@ -1,10 +1,10 @@
 from .classes import PipeEnd as end
-from .classes import PipeFactory as pipe
 from .classes import PipeObject as start
 from .classes import Tap as tap
 from .classes import TaskPipe as task
 from .classes import Then as then
 from .classes import WaitFor as wait
+from .classes import pipe
 
 __all__ = [
     "end",