+++ pipe_operator/python_flow/utils.py
@@ -1,7 +1,7 @@
 import inspect
 from typing import Any, Callable, Coroutine
 
-from typing_extensions import TypeGuard
+from typing_extensions import TypeIs
 
 from pipe_operator.python_flow.types import (
     AsyncCallable,
@@ -15,18 +15,20 @@ from pipe_operator.python_flow.types import (
 
 def is_async_pipeable(
     f: PipeableCallable[TInput, FuncParams, TOutput],
-) -> TypeGuard[AsyncCallable[TInput, FuncParams, TOutput]]:
-    """Checks if a function is "pipeable asynchronous" and provides a TypeGuard for it."""
+) -> TypeIs[AsyncCallable[TInput, FuncParams, TOutput]]:
+    """Checks if a function is "pipeable asynchronous" and provides a TypeIs for it."""
     return _is_async_function(f)
 
 
 def is_sync_pipeable(
     f: PipeableCallable[TInput, FuncParams, TOutput],
-) -> TypeGuard[SyncCallable[TInput, FuncParams, TOutput]]:
-    """Checks if a pipeable function is "pipeable synchronous" and provides a TypeGuard for it."""
+) -> TypeIs[SyncCallable[TInput, FuncParams, TOutput]]:
+    """Checks if a pipeable function is "pipeable synchronous" and provides a TypeIs for it."""
     return not _is_async_function(f)
 
 
-def _is_async_function(f: Callable[..., Any]) -> TypeGuard[Coroutine]:
-    """Checks if a function is asynchronous and provides a TypeGuard for it."""
+def _is_async_function(
+    f: Callable[..., Any],
+) -> TypeIs[Callable[..., Coroutine[Any, Any, Any]]]:
+    """Checks if a function is asynchronous and provides a TypeIs for it."""
     return inspect.iscoroutinefunction(f)